chore: shared prebundle
