+++ packages/plugin-babel/package.json
@@ -41,7 +41,7 @@
     "@scripts/test-helper": "workspace:*",
     "@types/node": "18.x",
     "babel-loader": "9.1.3",
-    "prebundle": "0.0.4",
+    "prebundle": "1.0.1",
     "typescript": "^5.4.2"
   },
   "peerDependencies": {
