+++ packages/plugin-rem/package.json
@@ -37,7 +37,7 @@
     "@scripts/test-helper": "workspace:*",
     "html-webpack-plugin": "npm:html-rspack-plugin@5.6.2",
     "postcss-pxtorem": "6.1.0",
-    "prebundle": "0.0.4",
+    "prebundle": "1.0.1",
     "typescript": "^5.4.2"
   },
   "peerDependencies": {
