+++ packages/shared/package.json
@@ -117,6 +117,7 @@
   "scripts": {
     "build": "modern build",
     "dev": "modern build --watch",
+    "prebundle": "prebundle",
     "test": "vitest run",
     "test:ui": "vitest --ui",
     "test:watch": "vitest dev --no-coverage"
@@ -127,12 +128,54 @@
     "postcss": "^8.4.38"
   },
   "devDependencies": {
+    "@types/connect": "3.4.38",
+    "@types/fs-extra": "^11.0.4",
+    "@types/less": "3.0.6",
     "@types/node": "18.x",
+    "@types/sass-loader": "8.0.8",
+    "@types/webpack-bundle-analyzer": "4.7.0",
+    "autoprefixer": "10.4.19",
+    "browserslist": "4.23.0",
+    "chokidar": "3.6.0",
+    "connect": "3.7.0",
+    "css-loader": "7.1.1",
+    "deepmerge": "^4.3.1",
+    "fs-extra": "^11.2.0",
+    "gzip-size": "^6.0.0",
     "html-webpack-plugin": "npm:html-rspack-plugin@5.6.2",
+    "http-proxy-middleware": "^2.0.6",
+    "icss-utils": "5.1.0",
+    "jiti": "^1.21.0",
+    "json5": "^2.2.3",
+    "less": "4.2.0",
+    "less-loader": "12.2.0",
+    "line-diff": "2.1.1",
+    "loader-utils2": "npm:loader-utils@2.0.4",
+    "mime-types": "^2.1.35",
     "mini-css-extract-plugin": "2.8.1",
+    "picocolors": "1.0.0",
+    "postcss-load-config": "5.0.3",
+    "postcss-loader": "8.1.1",
+    "postcss-modules-extract-imports": "3.1.0",
+    "postcss-modules-local-by-default": "4.0.5",
+    "postcss-modules-scope": "3.2.0",
+    "postcss-modules-values": "4.0.0",
+    "postcss-value-parser": "4.2.0",
+    "prebundle": "1.0.1",
+    "resolve-url-loader": "5.0.0",
+    "rslog": "^1.2.1",
+    "sass": "^1.75.0",
+    "sass-loader": "14.2.0",
+    "semver": "^7.6.0",
+    "style-loader": "3.3.4",
     "terser": "5.30.3",
     "typescript": "^5.4.2",
-    "webpack": "^5.91.0"
+    "webpack": "^5.91.0",
+    "webpack-bundle-analyzer": "^4.10.1",
+    "webpack-chain": "npm:webpack-5-chain@8.0.1",
+    "webpack-dev-middleware": "7.2.1",
+    "webpack-merge": "5.10.0",
+    "yaml": "^2.4.1"
   },
   "publishConfig": {
     "access": "public",
