--- scripts/prebundle/package.json
@@ -1,63 +1,0 @@
-{
-  "name": "@rsbuild/prebundle",
-  "private": true,
-  "version": "0.6.6",
-  "main": "./dist/index.js",
-  "scripts": {
-    "dev": "tsc --watch",
-    "build": "tsc",
-    "start": "pnpm build && node ./dist/index.js"
-  },
-  "dependencies": {
-    "@babel/core": "^7.24.4",
-    "@types/node": "18.x",
-    "@vercel/ncc": "0.38.1",
-    "dts-packer": "0.0.3",
-    "fast-glob": "^3.3.2",
-    "postcss": "^8.4.38",
-    "typescript": "^5.4.2"
-  },
-  "devDependencies": {
-    "@types/connect": "3.4.38",
-    "@types/fs-extra": "^11.0.4",
-    "@types/less": "3.0.6",
-    "@types/sass-loader": "8.0.8",
-    "@types/webpack-bundle-analyzer": "4.7.0",
-    "autoprefixer": "10.4.19",
-    "browserslist": "4.23.0",
-    "chokidar": "3.6.0",
-    "connect": "3.7.0",
-    "css-loader": "7.1.1",
-    "deepmerge": "^4.3.1",
-    "fs-extra": "^11.2.0",
-    "gzip-size": "^6.0.0",
-    "http-proxy-middleware": "^2.0.6",
-    "icss-utils": "5.1.0",
-    "jiti": "^1.21.0",
-    "json5": "^2.2.3",
-    "less": "4.2.0",
-    "less-loader": "12.2.0",
-    "line-diff": "2.1.1",
-    "loader-utils2": "npm:loader-utils@2.0.4",
-    "mime-types": "^2.1.35",
-    "picocolors": "1.0.0",
-    "postcss-load-config": "5.0.3",
-    "postcss-loader": "8.1.1",
-    "postcss-modules-extract-imports": "3.1.0",
-    "postcss-modules-local-by-default": "4.0.5",
-    "postcss-modules-scope": "3.2.0",
-    "postcss-modules-values": "4.0.0",
-    "postcss-value-parser": "4.2.0",
-    "resolve-url-loader": "5.0.0",
-    "rslog": "^1.2.1",
-    "sass": "^1.75.0",
-    "sass-loader": "14.2.0",
-    "semver": "^7.6.0",
-    "style-loader": "3.3.4",
-    "webpack-bundle-analyzer": "^4.10.1",
-    "webpack-chain": "npm:webpack-5-chain@8.0.1",
-    "webpack-dev-middleware": "7.2.1",
-    "webpack-merge": "5.10.0",
-    "yaml": "^2.4.1"
-  }
-}
