--- scripts/prebundle/src/global.d.ts
@@ -1,1 +1,0 @@
-declare module '@vercel/ncc';
