--- scripts/prebundle/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@rsbuild/tsconfig/base",
-  "compilerOptions": {
-    "baseUrl": "./",
-    "outDir": "./dist",
-    "module": "CommonJS",
-    "moduleResolution": "Node"
-  },
-  "include": ["src"]
-}
